refactor(Routes & Action creators):  Work under construction 🚧
- ApplicantActions file added
- GeneralActions fiel added
- ApplicantActions file imports updated
- GeneralActions fiel imports updated
