+++ client/src/components/Notifications.tsx
@@ -1,7 +1,7 @@
 import { FC, useEffect, useState, useRef } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import styled, { css } from "styled-components";
-import { getNotifications } from "../redux/actions/actionCreators";
+import { getNotifications } from "../redux/actions/private/generalActions";
 
 type Props = {
     role: string;
